requirements.txt
Django==4.2.6
django-bootstrap5==23.3
fontawesomefree==6.4.2 # Makes static files (css, fonts) available locally
django-fontawesome-6==1.0.0.0 # Provides an icon field for models and forms as well as handy shortcuts to render icons
django-split-settings==1.2.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
django-simple-history==3.4.0
django-registration-redux==2.12
django-debug-toolbar==4.2.0
django-bootstrap-datepicker-plus==5.0.4
django-tex==1.1.10
django-csp==3.7
django-compressor==4.4
django-libsass==0.9
django-betterforms==2.0.0
mysqlclient==2.2.0  # for production deployment
tzdata==2023.3

# Documentation
sphinxcontrib-django==2.3
sphinxcontrib-apidoc==0.3.0
recommonmark==0.7.1
django-docs==0.3.3
sphinx-rtd-theme==1.2.2
sphinx==6.2.1